Major update functional-wise, now servers static files
A number of helper functions have been added to print common headers and
error pages. A simple filename -> mime type mapper. In addition to this, if
the requested file actually exists and is readable (and is not the main
binary), it will serve the file.

In the case that the requested url is not actually a file, it will be routed
through the binary. This is not implemented yet. Instead, there is a simple
333 error page (no idea if that is a real error code)